+++ core/Cargo.toml
@@ -10,13 +10,16 @@ repository.workspace = true
 license.workspace = true
 
 [dependencies]
+ahash = { workspace = true }
 anyhow = { workspace = true }
 arc-swap = { workspace = true }
 async-trait = { workspace = true }
 bytes = { workspace = true, features = ["serde"] }
 bytesize = { workspace = true }
+castaway = { workspace = true }
 everscale-types = { workspace = true, features = ["blake3", "rayon"] }
 futures-util = { workspace = true }
+governor = { workspace = true }
 humantime = { workspace = true }
 metrics = { workspace = true }
 parking_lot = { workspace = true }
