+++ core/src/blockchain_rpc/service/mod.rs
@@ -1,21 +1,20 @@
+mod handlers;
 mod util;
 
-use std::num::{NonZeroU32, NonZeroU64};
+use std::num::NonZeroU32;
 use std::sync::Arc;
 
-use anyhow::Context;
 use bytes::{Buf, Bytes};
-use everscale_types::models::BlockId;
 use futures_util::Future;
 use metrics::Label;
 use serde::{Deserialize, Serialize};
 use tycho_block_util::message::validate_external_message;
 use tycho_network::{try_handle_prefix, InboundRequestMeta, Response, Service, ServiceRequest};
-use tycho_storage::{ArchiveId, BlockConnection, KeyBlocksDirection, PersistentStateKind, Storage};
+use tycho_storage::Storage;
 use tycho_util::futures::BoxFutureOrNoop;
 use tycho_util::metrics::HistogramGuard;
 
-use crate::blockchain_rpc::{BAD_REQUEST_ERROR_CODE, INTERNAL_ERROR_CODE, NOT_FOUND_ERROR_CODE};
+use crate::blockchain_rpc::service::util::{Constructor, RateLimiter};
 use crate::proto::blockchain::*;
 use crate::proto::overlay;
 
@@ -29,6 +28,10 @@ pub trait BroadcastListener: Send + Sync + 'static {
         meta: Arc<InboundRequestMeta>,
         message: Bytes,
     ) -> Self::HandleMessageFut<'_>;
+
+    fn is_noop(&self) -> bool {
+        false
+    }
 }
 
 #[derive(Debug, Default, Clone, Copy, Eq, PartialEq)]
@@ -40,6 +43,11 @@ impl BroadcastListener for NoopBroadcastListener {
     #[inline]
     fn handle_message(&self, _: Arc<InboundRequestMeta>, _: Bytes) -> Self::HandleMessageFut<'_> {
         futures_util::future::ready(())
+    }
+
+    #[inline]
+    fn is_noop(&self) -> bool {
+        true
     }
 }
 
@@ -56,6 +64,8 @@ pub struct BlockchainRpcServiceConfig {
     ///
     /// Default: yes.
     pub serve_persistent_states: bool,
+
+    pub rate_limits: RateLimits,
 }
 
 impl Default for BlockchainRpcServiceConfig {
@@ -63,6 +73,30 @@ impl Default for BlockchainRpcServiceConfig {
         Self {
             max_key_blocks_list_len: 8,
             serve_persistent_states: true,
+            rate_limits: Default::default(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+#[non_exhaustive]
+pub struct RateLimits {
+    /// rate limits for methods like `GetPersistentQueueStateInfo`, `GetArchiveInfo`, etc.
+    pub info_method_rps: NonZeroU32,
+    /// rate limits for methods like `GetPersistentQueueStateChunk`, `GetArchiveChunk`, etc.
+    pub chunk_method_rps: NonZeroU32,
+
+    /// Message broadcast rate limits
+    pub send_message: NonZeroU32,
+}
+
+impl Default for RateLimits {
+    fn default() -> Self {
+        Self {
+            info_method_rps: NonZeroU32::new(100).unwrap(),
+            chunk_method_rps: NonZeroU32::new(100).unwrap(),
+            send_message: NonZeroU32::new(10_000).unwrap(),
         }
     }
 }
@@ -82,6 +116,18 @@ where
         BlockchainRpcService {
             inner: Arc::new(Inner {
                 storage,
+                info_rate_limiter: RateLimiter::dashmap_with_hasher(
+                    governor::Quota::per_second(self.config.rate_limits.info_method_rps),
+                    Default::default(),
+                ),
+                chunk_rate_limiter: RateLimiter::dashmap_with_hasher(
+                    governor::Quota::per_second(self.config.rate_limits.chunk_method_rps),
+                    Default::default(),
+                ),
+                send_message_rate_limiter: RateLimiter::dashmap_with_hasher(
+                    governor::Quota::per_second(self.config.rate_limits.send_message),
+                    Default::default(),
+                ),
                 config: self.config,
                 broadcast_listener,
             }),
@@ -164,13 +210,20 @@ impl<B: BroadcastListener> Service<ServiceRequest> for BlockchainRpcService<B> {
             }
         };
 
-        let method = util::Constructor::from_tl_id(constructor);
+        let method = Constructor::from_tl_id(constructor);
         let label = vec![Label::new(
             "method",
             method.map_or("unknown", |m| m.as_str()),
         )];
-        let timer =
-            move || HistogramGuard::begin_with_labels_owned(RPC_METHOD_TIMINGS_METRIC, label);
+        let timer = {
+            let label = label.clone();
+            move || HistogramGuard::begin_with_labels_owned(RPC_METHOD_TIMINGS_METRIC, label)
+        };
+
+        if let Some(value) = self.inner.check_rate_limit(&req, method) {
+            metrics::counter!("tycho_rpc_rate_limit_exceeded_total", label).increment(1);
+            return value;
+        }
 
         let inner = self.inner.clone();
 
@@ -272,7 +325,20 @@ impl<B: BroadcastListener> Service<ServiceRequest> for BlockchainRpcService<B> {
     fn on_message(&self, mut req: ServiceRequest) -> Self::OnMessageFuture {
         use tl_proto::{BytesMeta, TlRead};
 
-        // TODO: Do nothing if `B` is `NoopBroadcastListener` via `castaway` ?
+        if self.inner.broadcast_listener.is_noop() {
+            return BoxFutureOrNoop::Noop;
+        }
+
+        if self
+            .inner
+            .send_message_rate_limiter
+            .check_key(&req.metadata.peer_id)
+            .is_err()
+        {
+            metrics::counter!("tycho_rpc_rate_limit_exceeded_total", "method" => "sendMessage")
+                .increment(1);
+            return BoxFutureOrNoop::Noop;
+        }
 
         // Require message body to contain at least two constructors.
         if req.body.len() < 8 {
@@ -344,6 +410,9 @@ struct Inner<B> {
     storage: Storage,
     config: BlockchainRpcServiceConfig,
     broadcast_listener: B,
+    info_rate_limiter: RateLimiter,
+    chunk_rate_limiter: RateLimiter,
+    send_message_rate_limiter: RateLimiter,
 }
 
 impl<B> Inner<B> {
@@ -351,321 +420,34 @@ impl<B> Inner<B> {
         &self.storage
     }
 
-    fn handle_get_next_key_block_ids(
+    fn check_rate_limit(
         &self,
-        req: &rpc::GetNextKeyBlockIds,
-    ) -> overlay::Response<KeyBlockIds> {
-        let block_handle_storage = self.storage().block_handle_storage();
-
-        let limit = std::cmp::min(req.max_size as usize, self.config.max_key_blocks_list_len);
-
-        let get_next_key_block_ids = || {
-            if !req.block_id.shard.is_masterchain() {
-                anyhow::bail!("first block id is not from masterchain");
-            }
-
-            let mut iterator = block_handle_storage
-                .key_blocks_iterator(KeyBlocksDirection::ForwardFrom(req.block_id.seqno))
-                .take(limit + 1);
-
-            if let Some(id) = iterator.next() {
-                anyhow::ensure!(
-                    id.root_hash == req.block_id.root_hash,
-                    "first block root hash mismatch"
-                );
-                anyhow::ensure!(
-                    id.file_hash == req.block_id.file_hash,
-                    "first block file hash mismatch"
-                );
-            }
-
-            Ok::<_, anyhow::Error>(iterator.take(limit).collect::<Vec<_>>())
+        req: &ServiceRequest,
+        method: Option<Constructor>,
+    ) -> Option<BoxFutureOrNoop<Option<Response>>> {
+        let rate_limiter = match method {
+            Some(
+                Constructor::GetPersistentQueueStateChunk
+                | Constructor::GetArchiveChunk
+                | Constructor::GetPersistentShardStateChunk
+                | Constructor::GetBlockDataChunk
+                | Constructor::GetNextBlockFull
+                | Constructor::GetBlockFull
+                | Constructor::GetNextKeyBlockIds
+                | Constructor::GetKeyBlockProof,
+            ) => &self.chunk_rate_limiter,
+            Some(
+                Constructor::GetPersistentShardStateInfo
+                | Constructor::GetPersistentQueueStateInfo
+                | Constructor::GetArchiveInfo
+                | Constructor::Ping,
+            ) => &self.info_rate_limiter,
+            None => return Some(BoxFutureOrNoop::Noop),
         };
 
-        match get_next_key_block_ids() {
-            Ok(ids) => {
-                let incomplete = ids.len() < limit;
-                overlay::Response::Ok(KeyBlockIds {
-                    block_ids: ids,
-                    incomplete,
-                })
-            }
-            Err(e) => {
-                tracing::warn!("get_next_key_block_ids failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    async fn handle_get_block_full(&self, req: &rpc::GetBlockFull) -> overlay::Response<BlockFull> {
-        match self.get_block_full(&req.block_id).await {
-            Ok(block_full) => overlay::Response::Ok(block_full),
-            Err(e) => {
-                tracing::warn!("get_block_full failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    async fn handle_get_next_block_full(
-        &self,
-        req: &rpc::GetNextBlockFull,
-    ) -> overlay::Response<BlockFull> {
-        let block_handle_storage = self.storage().block_handle_storage();
-        let block_connection_storage = self.storage().block_connection_storage();
-
-        let get_next_block_full = async {
-            let next_block_id = match block_handle_storage.load_handle(&req.prev_block_id) {
-                Some(handle) if handle.has_next1() => block_connection_storage
-                    .load_connection(&req.prev_block_id, BlockConnection::Next1)
-                    .context("connection not found")?,
-                _ => return Ok(BlockFull::NotFound),
-            };
-
-            self.get_block_full(&next_block_id).await
-        };
-
-        match get_next_block_full.await {
-            Ok(block_full) => overlay::Response::Ok(block_full),
-            Err(e) => {
-                tracing::warn!("get_next_block_full failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    fn handle_get_block_data_chunk(&self, req: &rpc::GetBlockDataChunk) -> overlay::Response<Data> {
-        let block_storage = self.storage.block_storage();
-        match block_storage.get_block_data_chunk(&req.block_id, req.offset) {
-            Ok(Some(data)) => overlay::Response::Ok(Data {
-                data: Bytes::from_owner(data),
-            }),
-            Ok(None) => overlay::Response::Err(NOT_FOUND_ERROR_CODE),
-            Err(e) => {
-                tracing::warn!("get_block_data_chunk failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    async fn handle_get_key_block_proof(
-        &self,
-        req: &rpc::GetKeyBlockProof,
-    ) -> overlay::Response<KeyBlockProof> {
-        let block_handle_storage = self.storage().block_handle_storage();
-        let block_storage = self.storage().block_storage();
-
-        let get_key_block_proof = async {
-            match block_handle_storage.load_handle(&req.block_id) {
-                Some(handle) if handle.has_proof() => {
-                    let data = block_storage.load_block_proof_raw(&handle).await?;
-                    Ok::<_, anyhow::Error>(KeyBlockProof::Found {
-                        proof: Bytes::from_owner(data),
-                    })
-                }
-                _ => Ok(KeyBlockProof::NotFound),
-            }
-        };
-
-        match get_key_block_proof.await {
-            Ok(key_block_proof) => overlay::Response::Ok(key_block_proof),
-            Err(e) => {
-                tracing::warn!("get_key_block_proof failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    async fn handle_get_archive_info(
-        &self,
-        req: &rpc::GetArchiveInfo,
-    ) -> overlay::Response<ArchiveInfo> {
-        let mc_seqno = req.mc_seqno;
-        let node_state = self.storage.node_state();
-
-        match node_state.load_last_mc_block_id() {
-            Some(last_applied_mc_block) => {
-                if mc_seqno > last_applied_mc_block.seqno {
-                    return overlay::Response::Ok(ArchiveInfo::TooNew);
-                }
-
-                let block_storage = self.storage().block_storage();
-
-                let id = block_storage.get_archive_id(mc_seqno);
-                let size_res = match id {
-                    ArchiveId::Found(id) => block_storage.get_archive_size(id),
-                    ArchiveId::TooNew | ArchiveId::NotFound => Ok(None),
-                };
-
-                overlay::Response::Ok(match (id, size_res) {
-                    (ArchiveId::Found(id), Ok(Some(size))) if size > 0 => ArchiveInfo::Found {
-                        id: id as u64,
-                        size: NonZeroU64::new(size as _).unwrap(),
-                        chunk_size: block_storage.archive_chunk_size(),
-                    },
-                    (ArchiveId::TooNew, Ok(None)) => ArchiveInfo::TooNew,
-                    _ => ArchiveInfo::NotFound,
-                })
-            }
-            None => {
-                tracing::warn!("get_archive_id failed: no blocks applied");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    async fn handle_get_archive_chunk(
-        &self,
-        req: &rpc::GetArchiveChunk,
-    ) -> overlay::Response<Data> {
-        let block_storage = self.storage.block_storage();
-
-        let get_archive_chunk = || async {
-            let archive_slice = block_storage
-                .get_archive_chunk(req.archive_id as u32, req.offset)
-                .await?;
-
-            Ok::<_, anyhow::Error>(archive_slice)
-        };
-
-        match get_archive_chunk().await {
-            Ok(data) => overlay::Response::Ok(Data {
-                data: Bytes::from_owner(data),
-            }),
-            Err(e) => {
-                tracing::warn!("get_archive_chunk failed: {e:?}");
-                overlay::Response::Err(INTERNAL_ERROR_CODE)
-            }
-        }
-    }
-
-    fn handle_get_persistent_state_info(
-        &self,
-        req: &rpc::GetPersistentShardStateInfo,
-    ) -> overlay::Response<PersistentStateInfo> {
-        let label = [("method", "getPersistentStateInfo")];
-        let _hist = HistogramGuard::begin_with_labels(RPC_METHOD_TIMINGS_METRIC, &label);
-        let res = self.read_persistent_state_info(&req.block_id, PersistentStateKind::Shard);
-        overlay::Response::Ok(res)
-    }
-
-    fn handle_get_queue_persistent_state_info(
-        &self,
-        req: &rpc::GetPersistentQueueStateInfo,
-    ) -> overlay::Response<PersistentStateInfo> {
-        let res = self.read_persistent_state_info(&req.block_id, PersistentStateKind::Queue);
-        overlay::Response::Ok(res)
-    }
-
-    async fn handle_get_persistent_shard_state_chunk(
-        &self,
-        req: &rpc::GetPersistentShardStateChunk,
-    ) -> overlay::Response<Data> {
-        self.read_persistent_state_chunk(&req.block_id, req.offset, PersistentStateKind::Shard)
-            .await
-    }
-
-    async fn handle_get_persistent_queue_state_chunk(
-        &self,
-        req: &rpc::GetPersistentQueueStateChunk,
-    ) -> overlay::Response<Data> {
-        self.read_persistent_state_chunk(&req.block_id, req.offset, PersistentStateKind::Queue)
-            .await
-    }
-}
-
-impl<B> Inner<B> {
-    async fn get_block_full(&self, block_id: &BlockId) -> anyhow::Result<BlockFull> {
-        let block_handle_storage = self.storage().block_handle_storage();
-        let block_storage = self.storage().block_storage();
-
-        let handle = match block_handle_storage.load_handle(block_id) {
-            Some(handle) if handle.has_all_block_parts() => handle,
-            _ => return Ok(BlockFull::NotFound),
-        };
-
-        let Some(data) = block_storage.get_block_data_chunk(block_id, 0)? else {
-            return Ok(BlockFull::NotFound);
-        };
-
-        let data_chunk_size = block_storage.block_data_chunk_size();
-        let data_size = if data.len() < data_chunk_size.get() as usize {
-            // NOTE: Skip one RocksDB read for relatively small blocks
-            //       Average block size is 4KB, while the chunk size is 1MB.
-            data.len() as u32
-        } else {
-            match block_storage.get_block_data_size(block_id)? {
-                Some(size) => size,
-                None => return Ok(BlockFull::NotFound),
-            }
-        };
-
-        let block = BlockData {
-            data: Bytes::from_owner(data),
-            size: NonZeroU32::new(data_size).expect("shouldn't happen"),
-            chunk_size: data_chunk_size,
-        };
-
-        let (proof, queue_diff) = tokio::join!(
-            block_storage.load_block_proof_raw(&handle),
-            block_storage.load_queue_diff_raw(&handle)
-        );
-
-        Ok(BlockFull::Found {
-            block_id: *block_id,
-            block,
-            proof: Bytes::from_owner(proof?),
-            queue_diff: Bytes::from_owner(queue_diff?),
-        })
-    }
-
-    fn read_persistent_state_info(
-        &self,
-        block_id: &BlockId,
-        state_kind: PersistentStateKind,
-    ) -> PersistentStateInfo {
-        let persistent_state_storage = self.storage().persistent_state_storage();
-        if self.config.serve_persistent_states {
-            if let Some(info) = persistent_state_storage.get_state_info(block_id, state_kind) {
-                return PersistentStateInfo::Found {
-                    size: info.size,
-                    chunk_size: info.chunk_size,
-                };
-            }
-        }
-        PersistentStateInfo::NotFound
-    }
-
-    async fn read_persistent_state_chunk(
-        &self,
-        block_id: &BlockId,
-        offset: u64,
-        state_kind: PersistentStateKind,
-    ) -> overlay::Response<Data> {
-        let persistent_state_storage = self.storage().persistent_state_storage();
-
-        let persistent_state_request_validation = || {
-            anyhow::ensure!(
-                self.config.serve_persistent_states,
-                "persistent states are disabled"
-            );
-            Ok::<_, anyhow::Error>(())
-        };
-
-        if let Err(e) = persistent_state_request_validation() {
-            tracing::debug!("persistent state request validation failed: {e:?}");
-            return overlay::Response::Err(BAD_REQUEST_ERROR_CODE);
-        }
-
-        match persistent_state_storage
-            .read_state_part(block_id, offset, state_kind)
-            .await
-        {
-            Some(data) => overlay::Response::Ok(Data { data: data.into() }),
-            None => {
-                tracing::debug!("failed to read persistent state part");
-                overlay::Response::Err(NOT_FOUND_ERROR_CODE)
-            }
-        }
+        rate_limiter
+            .check_key(&req.metadata.peer_id)
+            .err()
+            .map(|_| BoxFutureOrNoop::Noop)
     }
 }
